+++ src/plugin/plugin-manager.js
@@ -30,6 +30,14 @@ export default class PluginManager {
    * @private
    */
   _plugins: Map<string, BasePlugin> = new Map();
+  /**
+   * Is disabled plugin map.
+   * Maps plugin's name to a boolean.
+   * false means the plugin is disable. true or plugin name doesn't exist in the map means the plugin is not disable.
+   * @type {Map}
+   * @private
+   */
+  _isDisabledPluginMap: Map<string, boolean> = new Map();
 
   /**
    * Writes the plugin in the registry.
@@ -82,12 +90,18 @@ export default class PluginManager {
       throw new Error(Error.Severity.RECOVERABLE, Error.Category.PLAYER, Error.Code.RUNTIME_ERROR_NOT_REGISTERED_PLUGIN, name);
     }
     let pluginClass = PluginManager._registry.get(name);
-    if (pluginClass && pluginClass.isValid() && !config.disable) {
+    if (typeof config.disable === "boolean") {
+      this._isDisabledPluginMap.set(name, config.disable);
+    }
+    const isDisablePlugin = !!this._isDisabledPluginMap.get(name);
+    const isValidPlugin = pluginClass ? pluginClass.isValid() : false;
+    if (pluginClass && isValidPlugin && !isDisablePlugin) {
       this._plugins.set(name, pluginClass.createPlugin(name, player, config));
+      this._isDisabledPluginMap.set(name, false);
       logger.debug(`Plugin <${name}> has been loaded`);
       return true;
     }
-    logger.debug(`Plugin <${name}> isn\'t loaded, isValid()=false, disabled=${config.disable}`);
+    logger.debug(`Plugin <${name}> isn\'t loaded, isValid()=${isValidPlugin.toString()}, disabled=${isDisablePlugin.toString()}`);
     return false;
   }